refactor(game-engine): split gameReducer by action category
Extract 22 sub-reducer functions into two category files:
- lifecycleReducers.ts (10 handlers): init, restart, template,
  animation, join, leave, assign, viewedRole, fillBots, markBots
- nightActionReducers.ts (12 handlers): startNight, advance,
  endNight, record, applyResolver, witchCtx, confirm, wolfKill,
  wolfRobot, audio, rejected, revealAck

gameReducer.ts becomes a thin compose shell (641 -> 152 lines)
with switch dispatching to imported handlers + inline trivial
cases (clear/set single-field actions).
